% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEdgeMatrix.R
\name{as.dist.edgematrix}
\alias{as.dist.edgematrix}
\title{as.dist.edgematrix}
\usage{
\method{as.dist}{edgematrix}(m, diag = FALSE, upper = FALSE)
}
\arguments{
\item{m}{An `edgematrix` object.}

\item{diag}{logical value indicating whether the diagonal of the distance matrix should be printed by print.dist.}

\item{upper}{logical value indicating whether the upper triangle of the distance matrix should be printed by print.dist.}
}
\value{
A \code{\link[stats]{dist}} object.
}
\description{
Convert an edge matrix to a \code{dist} object.
}
\note{
This method converts the otherwise sparse edge matrix into a dense \code{dist} object,
where any distances absent from the edge matrix are represented as \code{NA}.
}
