% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEdgeMatrix.R
\name{buildWijMatrix}
\alias{buildWijMatrix}
\alias{buildWijMatrix.CsparseMatrix}
\alias{buildWijMatrix.TsparseMatrix}
\title{buildWijMatrix}
\usage{
buildWijMatrix(x, perplexity)

\method{buildWijMatrix}{TsparseMatrix}(x, perplexity = 50)

\method{buildWijMatrix}{CsparseMatrix}(x, perplexity = 50)
}
\arguments{
\item{x}{A sparse matrix}

\item{perplexity}{Given perplexity.}
}
\value{
A sparse matrix
}
\description{
Rescale the weights in an edge matrix to match a given perplexity.
}

