% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscan}
\alias{dbscan}
\title{dbscan}
\usage{
dbscan(data, neighbors, eps, minPts = nrow(data) + 1)
}
\arguments{
\item{data}{Input data, where examples are columns.}

\item{neighbors}{An adjacency matrix of the type produced by \code{\link{randomProjectionTreeSearch}}}

\item{eps}{See \code{\link[dbscan]{optics}}.}

\item{minPts}{}
}
\value{
An \code{\link[dbscan]{dbscan_fast}} object.
}
\description{
An implementation of the dbscan algorithm.
}

