% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{lof}
\alias{lof}
\title{Local Outlier Factor Score}
\usage{
lof(edges)
}
\arguments{
\item{edges}{An edge matrix of the type produced by \code{\link{buildEdgeMatrix}}.}
}
\value{
A vector of LOF values for each data point.
}
\description{
Calculate the Local Outlier Factor (LOF) score for each data point given knowledge
of k-Nearest Neighbors.
}

