% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{lv_optics}
\alias{lv_optics}
\title{lv_optics}
\usage{
lv_optics(edges, neighbors, eps = Inf, minPts = nrow(neighbors), eps_cl, xi,
  verbose = getOption("verbose", TRUE))
}
\arguments{
\item{edges}{A weighted graph of the type produced by \code{\link{buildEdgeMatrix}}.}

\item{neighbors}{An adjacency matrix of the type produced by \code{\link{randomProjectionTreeSearch}}}

\item{eps}{See \code{\link[dbscan]{optics}}.}

\item{minPts}{See \code{\link[dbscan]{optics}}.}

\item{eps_cl}{See \code{\link[dbscan]{optics}}.}

\item{xi}{See \code{\link[dbscan]{optics}}.}

\item{verbose}{Vebosity level.}
}
\value{
An \code{\link[dbscan]{optics}} object.
}
\description{
Experimental implementation of the OPTICS algorithm.
}
\details{
This is a preliminary implementation of a variant of the OPTICS algorithm that attempts
to leverage the \code{largeVis} nearest-neighbor search.
}
\note{
Support for dbscan and optics are preliminary, and not fully tested for
correctness.

This is not the original OPTICS algorithm. In particular, the neighbor-search strategy in
OPTICS is not used, in favor of using the pre-calculated neighbor matrix produced incidentally by
`largeVis`.
}
\references{
Mihael Ankerst, Markus M. Breunig, Hans-Peter Kriegel, Jörg Sander (1999). OPTICS: Ordering Points To Identify the Clustering Structure. ACM SIGMOD international conference on Management of data. ACM Press. pp. 49–60.
}

