% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{optics}
\alias{optics}
\title{OPTICS}
\usage{
optics(data, neighbors, eps, minPts = nrow(data) + 1, eps_cl, xi)
}
\arguments{
\item{data}{Input data, where examples are columns.}

\item{neighbors}{An adjacency matrix of the type produced by \code{\link{randomProjectionTreeSearch}}}

\item{eps}{See \code{\link[dbscan]{optics}}.}

\item{minPts}{See \code{\link[dbscan]{optics}}.}

\item{eps_cl}{See \code{\link[dbscan]{optics}}.}

\item{xi}{See \code{\link[dbscan]{optics}}.}
}
\value{
An \code{\link[dbscan]{optics}} object.
}
\description{
An implementation of the OPTICS algorithm.
}

