% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prjectKNNs.R
\name{projectKNNs}
\alias{projectKNNs}
\title{Project a distance matrix (presumably of k-nearest neighbors) into a lower-dimensional
 space using a proportionality function and stochastic gradient descent.}
\usage{
projectKNNs(x, dim, sgd.batches = nrow(x) * 10000, neg.samples = 5,
  weight.pos.samples = TRUE, distance.function = "1 / (1 + (2 * (X)^2))",
  gamma = 7, verbose = TRUE)
}
\description{
@param x A sparse matrix in triplet form
 @param dim The number of dimensions for the projection space
 @param sgd.batches The number of edges to process during SGD; defaults to 10000 * the number of rows in x
 @param neg.samples The number of negative edges to sample for each positive edge
 @param distance.function A function mapping the distance between two vertices in the lower-dimensional space to the probability that they are kNN's of each other.
 @param gamma Hyperparameter used in the objective function.  See the papers for details
 @param weight.pos.samples Whether to sample positive edges according to their edge weights (the default) or multiply the edge-loss by the edge-weight in the objective function.
 @return A dense [nrow(x),dim] matrix of the coordinates projecting x into the lower-dimensional space.
1 / (1 + (X)^2)',
1 / (1 + exp(- (X)^2))',
}

