% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionTreeSearch.R
\name{randomProjectionTreeSearch}
\alias{randomProjectionTreeSearch}
\title{Find approximate k-Nearest Neighbors using random projection tree search.}
\usage{
randomProjectionTreeSearch(x, K = 5, n.trees = 2, tree.threshold = max(10,
  ncol(x)), max.iter = 2, verbose = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{K}{How many nearest neighbors to seek for each node}

\item{n.trees}{The number of trees to build}

\item{tree.threshold}{The threshold for creating a new branch.  The paper authors suggest
using a value equivalent to the number of features in the input set.}

\item{max.iter}{Number of iterations in the neighborhood exploration phase}

\item{verbose}{Whether to print verbose logging using the \code{progress} package}
}
\value{
A [K, N] matrix of the approximate K nearest neighbors for each vertex.
}
\description{
This is a very fast and accurate algorithm for finding k-nearest neighbors.
}
\examples{
\dontrun{
library(kknn) # Shamelessly borrowing sample data from another package
data(miete)
miete <- model.matrix(~ ., miete)
system.time(neighbors <- randomProjectionTreeSearch(miete))
}

}

