% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectKNNs.R
\name{sgdBatches}
\alias{sgdBatches}
\title{sgdBatches}
\usage{
sgdBatches(N, E = 150 * N/2)
}
\arguments{
\item{N}{Number of vertices.}

\item{E}{Number of edges.}
}
\value{
The recommended number of sgd batches.
}
\description{
Calculate the default number of batches for a given number of vertices and edges.
}
\details{
The formula used is the one used by the \code{LargeVis} reference implementation.  This is substantially less than the recommendation \eqn{E * 10000} in the original paper.
}
