% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{as.dendrogram.hdbscan}
\alias{as.dendrogram.hdbscan}
\title{as.dendrogram.hdbscan}
\usage{
\method{as.dendrogram}{hdbscan}(x)
}
\description{
Convert hdbscan object to dendrogram.
}
\examples{
data(iris)
neighbors <- randomProjectionTreeSearch(t(iris[,1:4]), K = 20)
edges <- buildEdgeMatrix(t(iris[,1:4]), neighbors)
hdbscanobj <- hdbscan(edges, minPts = 10, K = 5)
plot(as.dendrogram(hdbscanobj))
}

