% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEdgeMatrix.R
\name{buildEdgeMatrix}
\alias{buildEdgeMatrix}
\title{Build an nearest-neighbor graph weighted by distance.}
\usage{
buildEdgeMatrix(data, neighbors, distance_method = "Euclidean",
  verbose = TRUE)
}
\arguments{
\item{data}{A matrix with a number of columns equal to the number of columns in `x`}

\item{distance_method}{One of "Euclidean" or "Cosine"}

\item{verbose}{Verbosity}

\item{x}{An edge matrix, 0-indexed, where each vertex is a column}
}
\value{
A `sparseMatrix`
}
\description{
Build an nearest-neighbor graph weighted by distance.
}

