% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEdgeMatrix.R
\name{buildEdgeMatrix}
\alias{buildEdgeMatrix}
\title{Build an nearest-neighbor graph weighted by distance.}
\usage{
buildEdgeMatrix(data, neighbors, distance_method = "Euclidean",
  verbose = getOption("verbose", TRUE))
}
\arguments{
\item{data}{A matrix with a number of columns equal to the number of columns in `x`}

\item{neighbors}{An adjacency matrix of the type produced by \code{\link{randomProjectionTreeSearch}}.}

\item{distance_method}{One of "Euclidean" or "Cosine"}

\item{verbose}{Verbosity}
}
\value{
A `sparseMatrix`
}
\description{
Build an nearest-neighbor graph weighted by distance.
}

