% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{dbscan}
\alias{dbscan}
\title{dbscan}
\usage{
dbscan(edges, neighbors, eps = Inf, minPts = nrow(neighbors),
  verbose = getOption("verbose", TRUE))
}
\arguments{
\item{edges}{A weighted graph of the type produced by \code{\link{buildEdgeMatrix}}.}

\item{neighbors}{An adjacency matrix of the type produced by \code{\link{randomProjectionTreeSearch}}}

\item{eps}{See \code{\link[dbscan]{dbscan}}.}

\item{minPts}{See \code{\link[dbscan]{dbscan}}.}

\item{verbose}{Vebosity level.}
}
\value{
A \code{\link[dbscan]{dbscan}} object.
}
\description{
Experimental implementation of the DBSCAN algorithm.
}
\references{
Martin Ester, Hans-Peter Kriegel, Jörg Sander, Xiaowei Xu (1996). Evangelos Simoudis, Jiawei Han, Usama M. Fayyad, eds. A density-based algorithm for discovering clusters in large spatial databases with noise. Proceedings of the Second International Conference on Knowledge Discovery and Data Mining (KDD-96). AAAI Press. pp. 226–231. ISBN 1-57735-004-9.
}

