% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{hdbscan}
\alias{hdbscan}
\title{hdbscan}
\usage{
hdbscan(edges, minPts = 20, K = 5, verbose = getOption("verbose", TRUE))
}
\arguments{
\item{edges}{An edge matrix of the type returned by \code{\link{buildEdgeMatrix}}.}

\item{minPts}{The minimum number of points in a cluster.}

\item{K}{The number of points in the core neighborhood. (See details.)}

\item{verbose}{Verbosity.}
}
\value{
An object of type \code{hdbscan} with the following fields:
}
\description{
Experiment implemenation of the hdbscan algorithm.
}
\details{
The hyperparameter \code{K} controls the size of core neighborhoods. 
The algorithm measures the density around a point as 1 / the distance between
that point and its \code{K}th nearest neighbor. A low value of \code{K} is similar
to clustering nearest neighbors rather than based on density. A high value of
\code{K} may cause the algorithm to miss some (usually contrived) clustering
patterns, such as where clusters are made up of points arranged in lines to form
shapes.
}

