% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeVis.R
\name{largeVis}
\alias{largeVis}
\title{Apply the LargeVis algorithm for visualizing large high-dimensional datasets.}
\usage{
largeVis(x, dim = 2, K = 40, pca.first = TRUE, pca.dims = 50,
  n.trees = 2, tree.threshold = K * 2, max.iter = 2,
  distance.method = "euclidean", perplexity = K, sgd.batches = nrow(x) *
  10000, neg.samples = 5, weight.pos.samples = TRUE,
  distance.function = "1 / (1 + (2 * (X)^2))", gamma = 7, verbose = TRUE,
  ...)
}
\arguments{
\item{x}{A matrix}

\item{dim}{The number of dimensions in the output}

\item{K}{The number of nearest-neighbors to use in computing the graph}

\item{pca.first}{Whether to apply pca first (can speed-up distance calculations)}

\item{pca.dims}{How many pca dimensions to use}

\item{n.trees}{The number of random projections trees to create when approximating nearest neighbors}

\item{tree.threshold}{The threshold for creating separate projection tree branches}

\item{max.iter}{Maximum number of iterations during the neighbor search (the paper recommends 1)}

\item{distance.method}{The method used to compute distances when creating the kNN graph; passed to \code{proxy::dist()}}

\item{perplexity}{See paper}

\item{sgd.batches}{See `projectKNNs`}

\item{neg.samples}{See `projectKNNs`}

\item{weight.pos.samples}{See `projectKNNs`}

\item{distance.function}{See `projectKNNs`}

\item{gamma}{See `projectKNNs`}

\item{verbose}{Verbosity}

\item{...}{See paper}
}
\value{
A `largeVis` object with the following slots:
}
\description{
Apply the LargeVis algorithm for visualizing large high-dimensional datasets.
}
\examples{

}

