% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeVis.R, R/largeVisBackup.R
\name{largeVis}
\alias{largeVis}
\title{Apply the LargeVis algorithm for visualizing large high-dimensional datasets.}
\usage{
largeVis(x, dim = 2, K = 40, pca.first = TRUE, pca.dims = 50,
  n.trees = 2, tree.threshold = K * 2, max.iter = 2,
  distance.method = "euclidean", perplexity = K, sgd.batches = nrow(x) *
  10000, M = 5, weight.pos.samples = TRUE,
  distance.function = "1 / (1 + (2 * (X)^2))", gamma = 7, verbose = TRUE,
  ...)

largeVis(x, dim = 2, K = 40, pca.first = TRUE, pca.dims = 50,
  n.trees = 2, tree.threshold = K * 2, max.iter = 2,
  distance.method = "euclidean", perplexity = K, sgd.batches = nrow(x) *
  10000, M = 5, weight.pos.samples = TRUE,
  distance.function = "1 / (1 + (2 * (X)^2))", gamma = 7, verbose = TRUE,
  ...)
}
\arguments{
\item{x}{A matrix}

\item{dim}{The number of dimensions in the output}

\item{K}{The number of nearest-neighbors to use in computing the graph}

\item{pca.first}{Whether to apply pca first (can speed-up distance calculations)}

\item{pca.dims}{How many pca dimensions to use}

\item{n.trees}{See \link{randomProjectionTreeSearch}}

\item{tree.threshold}{See \link{randomProjectionTreeSearch}}

\item{max.iter}{See \link{randomProjectionTreeSearch}}

\item{distance.method}{See \link{randomProjectionTreeSearch}}

\item{perplexity}{See paper}

\item{sgd.batches}{See \link{projectKNNs}}

\item{M}{See \link{projectKNNs}}

\item{weight.pos.samples}{See \link{projectKNNs}}

\item{distance.function}{See \link{projectKNNs}}

\item{gamma}{See `projectKNNs`}

\item{verbose}{Verbosity}

\item{...}{See paper}

\item{alpha}{See \link{projectKNNs}}

\item{x}{A matrix}

\item{dim}{The number of dimensions in the output}

\item{K}{The number of nearest-neighbors to use in computing the graph}

\item{pca.first}{Whether to apply pca first (can speed-up distance calculations)}

\item{pca.dims}{How many pca dimensions to use}

\item{n.trees}{See \code\link{randomProjectionTreeSearch}}

\item{tree.threshold}{See \code\link{randomProjectionTreeSearch}}

\item{max.iter}{See \code\link{randomProjectionTreeSearch}}

\item{distance.method}{See \code\link{randomProjectionTreeSearch}}

\item{perplexity}{See paper}

\item{sgd.batches}{See \code\link{projectKNNs}}

\item{M}{See \code\link{projectKNNs}}

\item{weight.pos.samples}{See \code\link{projectKNNs}}

\item{distance.function}{See \code\link{projectKNNs}}

\item{gamma}{See `projectKNNs`}

\item{verbose}{Verbosity}

\item{...}{See paper}
}
\value{
A `largeVis` object with the following slots:

A `largeVis` object with the following slots:
}
\description{
See \href{https://arxiv.org/abs/1602.00370}{Original paper}

See \href{https://arxiv.org/abs/1602.00370}{Original paper}
}
\examples{


}

