% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionTreeSearch.R
\name{randomProjectionTreeSearch}
\alias{randomProjectionTreeSearch}
\alias{randomProjectionTreeSearch.matrix}
\alias{randomProjectionTreeSearch.CsparseMatrix}
\alias{randomProjectionTreeSearch.TsparseMatrix}
\alias{randomProjectionTreeSearchFromIndex}
\title{Find approximate k-Nearest Neighbors using random projection tree search.}
\usage{
randomProjectionTreeSearch(
  x,
  K = 150,
  n_trees = 50,
  max_iter = 1,
  distance_method = "Euclidean",
  save_file = NULL,
  verbose = getOption("verbose", TRUE)
)

\method{randomProjectionTreeSearch}{matrix}(
  x,
  K = 150,
  n_trees = 50,
  max_iter = 1,
  distance_method = "Euclidean",
  save_file = NULL,
  verbose = getOption("verbose", TRUE)
)

\method{randomProjectionTreeSearch}{CsparseMatrix}(
  x,
  K = 150,
  n_trees = 50,
  max_iter = 1,
  distance_method = "Euclidean",
  save_file = NULL,
  verbose = getOption("verbose", TRUE)
)

\method{randomProjectionTreeSearch}{TsparseMatrix}(
  x,
  K = 150,
  n_trees = 50,
  max_iter = 1,
  distance_method = "Euclidean",
  save_file = NULL,
  verbose = getOption("verbose", TRUE)
)

randomProjectionTreeSearchFromIndex(
  save_file,
  D,
  K = 150,
  max_iter = 1,
  distance_method = "Euclidean",
  verbose = getOption("verbose", TRUE)
)
}
\arguments{
\item{x}{A (potentially sparse) matrix, where examples are columnns and features are rows.}

\item{K}{The number of nearest neighbors to find}

\item{n_trees}{The number of trees to build.}

\item{max_iter}{Number of iterations in the neighborhood exploration phase.}

\item{distance_method}{One of "Euclidean" or "Cosine" - must match the previously saved index.}

\item{save_file}{The path to the index}

\item{verbose}{Whether to print verbose logging using the \code{progress} package.}

\item{D}{Dimensionality of the dataset}
}
\value{
A [K, N] matrix of the approximate K nearest neighbors for each vertex.

A [K, N] matrix of the approximate K nearest neighbors for each vertex.
}
\description{
A fast and accurate algorithm for finding approximate k-nearest neighbors.
}
\details{
Note that the algorithm does not guarantee that it will find K neighbors for each node. A
warning will be issued if it finds fewer neighbors than requested. If the input data contains
distinct partitionable clusters, try increasing the \code{tree_threshold} to increase the number
of returned neighbors.
}
