% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionTreeSearch.R
\name{randomProjectionTreeSearch}
\alias{randomProjectionTreeSearch}
\title{Find approximate k-Nearest Neighbors using random projection tree search}
\usage{
randomProjectionTreeSearch(x, K = 5, n.trees = 2, tree.threshold = K * 2,
  max.iter = 2, verbose = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{K}{How many nearest neighbors to seek for each node}

\item{n.trees}{The number of trees to build}

\item{tree.threshold}{The threshold for creating a new branch}

\item{max.iter}{Number of iterations in the neighborhood exploration phase}

\item{verbose}{Whether to print verbose logging using the \code{progress} package}
}
\value{
A [nrow(x), K] integer matrix showing the estimated K nearest neighbors for each vertex.
}
\description{
Find approximate k-Nearest Neighbors using random projection tree search
}
\examples{

}

