% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionTreeSearch.R
\name{randomProjectionTreeSearch}
\alias{randomProjectionTreeSearch}
\title{Find approximate k-Nearest Neighbors using random projection tree search.}
\usage{
randomProjectionTreeSearch(x, K = 5, n.trees = 2, tree.threshold = min(K *
  1.5, nrow(x)), max.iter = 2, verbose = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{K}{How many nearest neighbors to seek for each node}

\item{n.trees}{The number of trees to build}

\item{tree.threshold}{The threshold for creating a new branch.  A value of 1.5 * K guarantees that
the algorithm will return K candidate nearest neighbors for each vertex.}

\item{max.iter}{Number of iterations in the neighborhood exploration phase}

\item{verbose}{Whether to print verbose logging using the \code{progress} package}
}
\value{
A [N, K] integer matrix showing the estimated K nearest neighbors for each vertex.
}
\description{
This is a very fast and accurate algorithm for finding k-nearest neighbors.
}
\examples{

}

