% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionTreeSearch.R
\name{randomProjectionTreeSearch}
\alias{randomProjectionTreeSearch}
\title{Find approximate k-Nearest Neighbors using random projection tree search.}
\usage{
randomProjectionTreeSearch(x, K = 5, n_trees = 2, tree_threshold = max(10,
  ncol(x)), max_iter = 2, max_depth = 32, verbose = TRUE)
}
\arguments{
\item{x}{A matrix}

\item{K}{How many nearest neighbors to seek for each node}

\item{n_trees}{The number of trees to build}

\item{tree_threshold}{The threshold for creating a new branch.  The paper authors suggest
using a value equivalent to the number of features in the input set.}

\item{max_iter}{Number of iterations in the neighborhood exploration phase}

\item{max_depth}{The maximum level of recursion (relates to memory consumption during the tree search)}

\item{verbose}{Whether to print verbose logging using the \code{progress} package}
}
\value{
A [K, N] matrix of the approximate K nearest neighbors for each vertex.
}
\description{
This is a very fast and accurate algorithm for finding k-nearest neighbors.
}
\examples{
\dontrun{
library(kknn) # Shamelessly borrowing sample data from another package
data(miete)
miete <- model.matrix(~ ., miete)
system.time(neighbors <- randomProjectionTreeSearch(miete))
}

}

