% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectKNNs.R
\name{sgdBatches}
\alias{sgdBatches}
\title{sgdBatches}
\usage{
sgdBatches(N, E = 150 * N/2)
}
\arguments{
\item{N}{Number of vertices.}

\item{E}{Number of edges.}
}
\value{
The recommended number of sgd batches.
}
\description{
Calculate the default number of batches for a given number of vertices and edges.
}
\details{
The formula used is the one used by the \code{LargeVis} reference implementation.  This is substantially less than the recommendation \eqn{E * 10000} in the original paper.
}
\examples{
# Observe that increasing K has no effect on processing time
N <- 70000 # MNIST
K <- 10:250
plot(K, sgdBatches(rep(N, length(K)), N * K / 2))

# Observe that processing time scales linarly with N
N <- c(seq(from = 1, to = 10000, by = 100), seq(from = 10000, to = 10000000, by = 1000))
plot(N, sgdBatches(N))
}
